pinn: move modules
