+++ fealpy/pinn/__init__.py
@@ -1,4 +1,5 @@
 
 from .sampler import *
-from .machine import LearningMachine, Solution
-from .grad import gradient
+from .machine import LearningMachine
+from .grad import gradient, grad_by_fts, grad_of_fts
+from .tools import mkfs, use_mkfs
